+++ R/add_ci.R
@@ -73,7 +73,7 @@ add_ci <- function(x, ...) {
 #' @rdname add_ci
 #' @export
 add_ci.tbl_summary <- function(x,
-                               method = NULL,
+                               method = list(all_continuous() ~ "t.test", all_categorical() ~ "wilson"),
                                include = everything(),
                                statistic =
                                  list(all_continuous() ~ "{conf.low}, {conf.high}",
@@ -98,17 +98,9 @@ add_ci.tbl_summary <- function(x,
     include = {{ include }}
   )
 
-  default_method <- list(
-    rlang::new_formula(
-      rlang::expr(all_continuous()),
-      get_theme_element("add_ci.tbl_summary-attr:method.continuous", default = "t.test")
-    ),
-    rlang::new_formula(
-      rlang::expr(all_categorical()),
-      get_theme_element("add_ci.tbl_summary-attr:method.categorical", default = "wilson")
-    )
-  )
-  if (is.null(method)) method <- default_method
+  if (missing(method)) {
+    method <- get_theme_element("add_ci.tbl_summary-arg:method", default = list(all_continuous() ~ "t.test", all_categorical() ~ "wilson"))
+  }
 
   cards::process_formula_selectors(
     data = scope_table_body(x$table_body |> dplyr::filter(.data$variable %in% .env$include)),
@@ -118,7 +110,7 @@ add_ci.tbl_summary <- function(x,
   )
   cards::fill_formula_selectors(
     data = scope_table_body(x$table_body |> dplyr::filter(.data$variable %in% .env$include)),
-    method = default_method,
+    method = eval(formals(asNamespace("gtsummary")[["add_ci.tbl_summary"]])[["method"]]),
     statistic = eval(formals(asNamespace("gtsummary")[["add_ci.tbl_summary"]])[["statistic"]]),
     style_fun = eval(formals(asNamespace("gtsummary")[["add_ci.tbl_summary"]])[["style_fun"]])
   )