+++ R/add_ci.tbl_svysummary.R
@@ -44,7 +44,7 @@
 #'   modify_header(all_stat_cols() ~ "**{level}**") |>
 #'   modify_spanning_header(all_stat_cols() ~ "**Survived**")
 add_ci.tbl_svysummary <- function(x,
-                                  method = NULL,
+                                  method = list(all_continuous() ~ "svymean", all_categorical() ~ "svyprop.logit"),
                                   include = everything(),
                                   statistic =
                                     list(all_continuous() ~ "{conf.low}, {conf.high}",
@@ -74,17 +74,9 @@ add_ci.tbl_svysummary <- function(x,
     include = {{ include }}
   )
 
-  default_method <- list(
-    rlang::new_formula(
-      rlang::expr(all_continuous()),
-      get_theme_element("add_ci.tbl_svysummary-attr:method.continuous", default = "svymean")
-    ),
-    rlang::new_formula(
-      rlang::expr(all_categorical()),
-      get_theme_element("add_ci.tbl_svysummary-attr:method.categorical", default = "svyprop.logit")
-    )
-  )
-  if (is.null(method)) method <- default_method
+  if (missing(method)) {
+    method <- get_theme_element("add_ci.tbl_svysummary-arg:method", default = list(all_continuous() ~ "svymean", all_categorical() ~ "svyprop.logit"))
+  }
 
   cards::process_formula_selectors(
     data = scope_table_body(x$table_body |> dplyr::filter(.data$variable %in% .env$include)),
@@ -94,7 +86,7 @@ add_ci.tbl_svysummary <- function(x,
   )
   cards::fill_formula_selectors(
     data = scope_table_body(x$table_body |> dplyr::filter(.data$variable %in% .env$include)),
-    method = default_method,
+    method = eval(formals(asNamespace("gtsummary")[["add_ci.tbl_svysummary"]])[["method"]]),
     statistic = eval(formals(asNamespace("gtsummary")[["add_ci.tbl_svysummary"]])[["statistic"]]),
     style_fun = eval(formals(asNamespace("gtsummary")[["add_ci.tbl_svysummary"]])[["style_fun"]])
   )